Update reco_lab_web_page.json
The example didn't work any more because of changes in the visual-recognition node's output format (now named visual-recognition-v3) . I edited the "Report" template according to the new format and added a function node to format the output to see in debug the classes found by the visual-recognition node. Note that the images in http://visual-recognition-demo.mybluemix.net/images/*.jpg are not longer available.